+++ client/package.json
@@ -8,6 +8,7 @@
     "@fullcalendar/react": "^4.2.0",
     "axios": "^0.19.0",
     "css-loader": "^3.1.0",
+    "moment": "^2.24.0",
     "node-sass": "^4.12.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
