+++ client/src/components/private/Calendar.js
@@ -5,6 +5,7 @@ import { Button, Modal, ModalBody, Row, Card, CardHeader } from "reactstrap";
 import FullCalendar from '@fullcalendar/react'
 import dayGridPlugin from '@fullcalendar/daygrid'
 import interactionPlugin from '@fullcalendar/interaction';
+import moment from 'moment';
 import API from "../../utils/API";
 import './main.scss' // webpack must be configured to do this
 
@@ -13,58 +14,69 @@ export default class Calendar extends React.Component {
     super(props)
 
 
- this.calendarComponentRef = React.createRef()
-  this.state = {
-    events: 
-    [{
-     
-    },
-    {
-      title: "Repair Fence",
-      start: "2019-08-07",
-      end: "2019-08-09",
-      editable: true
-    },
-    {
-      title: "Change Air Filter",
-      start: "2019-08-09",
-      end: "2019-08-09",
-      editable: true
-    },
-  ],
-    search: "",
-    eventToSave: {
-      repairType: "",
-      title: "",
-      startDate: "", // need to parse before putting into events array
-      endDate: "",
-      startTime: "",
-      endTime: "",
-      cost: 0,
-      priority: "",
-      status: "",
-      isVendor: false,
-      vendor: "",
-      notes: "",
-      editable: true,
-    },
-    modal: false,
-  };
+    this.calendarComponentRef = React.createRef()
+    this.state = {
+        events: [],
+    //   search: "",
+        repairType: "",
+        title: "",
+        startDate: "", // need to parse before putting into events array
+        endDate: "",
+        startTime: "",
+        endTime: "",
+        cost: 0,
+        priority: "",
+        status: "",
+        isVendor: false,
+        vendor: "",
+        notes: "",
+        editable: true,
+        
+        modal: false,
+        // detailModal: false,
+    };
 
- this.handleFormSubmit = this.handleFormSubmit.bind(this);
- this.handleInputChange = this.handleInputChange.bind(this);
-} 
+    this.handleFormSubmit = this.handleFormSubmit.bind(this);
+    this.handleInputChange = this.handleInputChange.bind(this);
+    } 
   
-     /*  this.handleSaveBook = this.handleSaveBook.bind(this);
-     
-      
-    } */
+    componentDidMount() {
+        this.loadEvents();
+    }
+
+    loadEvents = () => { 
+        
+        API.getEvents()
+        .then(events => {
+    //        console.log(books);
+            const eventList =  events.data.map(item => ({
+                title: item.title,
+                start: item.start,
+                end: item.end,
+                editable: item.editable 
+            }))      
+        /*  console.log("results of get in getBooks: " + JSON.stringify(books));  // working */
+            this.setState({ 
+            events: eventList,
+            });
+        })
+        .catch(err => console.log(err));       
+        };
+
     
     toggle = () => {
       this.setState(prevState => ({
         modal: !prevState.modal,
       }));
     };
+
+    handleDateClick = (dateObj) => {
+        const {date} = dateObj;
+        const newEvents = this.state.events.filter (event => event.start === date);
+        console.log("this.state.events: " + this.state.events);
+        console.log("Date: " + date);
+        console.log(newEvents);
+    }
    
     handleNewEvent = (arg) => {                            // handleDateClick = 
   //    console.log(this.props.user_id);  // working
@@ -72,87 +84,102 @@ export default class Calendar extends React.Component {
         modal: !prevState.modal,
       })
       );
-      // eslint-disable-next-line no-restricted-globals
-    //  if (confirm('Would you like to add an event to ' + arg.dateStr + ' ?')) {  // Need a form in a modal here.
-     /*    console.log(this.state.events);
-        this.setState({  // add new event data
-          events: this.state.events.concat({ // creates a new array
-            title: arg.title,
-            start: arg.start,
-            end: arg.end,
-            editable: true
-          }) */
-  //      })
-    //  }
     }
+
 
     handleCheck = () => {
       this.setState({ isVendor: !this.state.isVendor });
     };
 
+
     handleInputChange = event => {
       // Getting the value and name of the input which triggered the change
       const { name, value } = event.target;
-        
+           console.log(event.target);
       // Updating the input's state
       this.setState({
-        eventToSave: {
+       
         [name]: value,
-        }
+      
       });
-      console.log(this.state.eventToSave);  // Not working for dropdowns?
     };
+
+  
 
     handleFormSubmit = event => {
       // Preventing the default behavior of the form submit (which is to refresh the page)
       event.preventDefault();
-      console.log(this.state.eventToSave);
-      let newEvent = {
-        repairType: this.state.eventToSave.repairType,
-        title: this.state.eventToSave.title,
-        cost: this.state.eventToSave.cost,
-        priority: this.state.eventToSave.priority,
-        status: this.state.eventToSave.status,
-        startDate: this.state.eventToSave.startDate,
-        endDate: this.state.eventToSave.endDate,
-        startTime: this.state.eventToSave.startTime,
-        endTime: this.state.eventToSave.endTime,
-        isVendor: this.state.eventToSave.isVendor,
-        vendor: this.state.eventToSave.vendor,
-        notes: this.state.eventToSave,            // .notes
-        editable: true,                         // This is the only thing being pushed to the DB.
+
+      //   Sample:  2018-06-01T12:30:00
+      let parsedStartDate = this.state.startDate + "T" + this.state.startTime + ":00";
+      let parsedEndDate = this.state.endDate + "T" + this.state.endTime + ":00";
+
+      let momentDate = moment().format(parsedStartDate);
+      console.log(momentDate);
+      console.log(typeof momentDate);
+   
+      let newDateDate = new Date(parsedStartDate);
+      console.log("newDateDate: " + newDateDate)
+
+      
+      let newRepair = {
+        repairType: this.state.repairType,
+        title: this.state.title,
+        cost: this.state.cost,
+        priority: this.state.priority,
+        status: this.state.status,
+        startDate: parsedStartDate,
+        endDate: parsedEndDate,
+        vendor: this.state.vendor,
+        notes: this.state.notes,
       }
-      API.saveRepair(newEvent)
-          .then(newEvent => {
-            console.log(newEvent);
+
+      let newEvent ={
+        title: this.state.title,
+        start: parsedStartDate,
+        end: parsedEndDate,
+        editable: true
+
+      }
+      API.saveRepair(newRepair)
+          .then(newRepair => {
+            console.log(newRepair);
           })
           .catch(err => console.log(err));
 
       API.saveEvent(newEvent)
       .then(newEvent => {
-
-        this.setState({ 
-          events: this.state.events.concat({ // creates a new array
-            title: newEvent.title,
-            start: newEvent.start,
-            end: newEvent.end,
-            editable: true
-          })
-        });
         this.toggle();
       })
       .catch(err => console.log(err));
-      
-      
-      
-      
+
+      API.getEvents()
+      .then(response => {
+    //    console.log(response);
+        this.setState({ 
+            events: response.data
+          });
+      })
+      .catch(err => console.log(err));
+
+      this.setState({
+        repairType: "",
+        title: "",
+        startDate: "", // need to parse before putting into events array
+        endDate: "",
+        startTime: "",
+        endTime: "",
+        cost: 0,
+        priority: "",
+        status: "",
+        isVendor: false,
+        vendor: "",
+        notes: "",
+        editable: true,
+      })
+
+
     };
-
-/*   gotoPast = () => {
-    let calendarApi = this.calendarComponentRef.current.getApi()
-    calendarApi.gotoDate('2000-01-01') // call a method on the Calendar object
-  } */
-
 
 
 /*   loadEvents = (start, end, timezone, callback) => {
@@ -180,17 +207,15 @@ export default class Calendar extends React.Component {
 render() {
   return (
       <div>
-      {/* <h3>Click a date/time to add an event</h3> */}
       <button onClick={ this.handleNewEvent }><h3>Add Event</h3></button>&nbsp;
       <div className='calendar-container'>
         <div className='calendar-top'>
-          
         </div>
         <div className='calendar'></div>
     <FullCalendar
-    timeZone= "UTC"
     defaultView="dayGridMonth"
-    /* dateClick={this.handleDateClick} */
+    dateClick={this.handleDateClick }
+    /* this.handleDateClick */
     
     header={{
               left: 'prev,next today',
@@ -200,6 +225,7 @@ render() {
     plugins={[ dayGridPlugin, interactionPlugin ]}
     ref={ this.calendarComponentRef }
     events={ this.state.events }
+
     >
 
     </FullCalendar>
@@ -225,7 +251,7 @@ render() {
                   <select
                     className="form-control"
                     id="repairTypeSelect"
-                    value={this.state.eventToSave.repairType}
+                    value={this.state.repairType}
                     name="repairType"
                     onChange={this.handleInputChange}
                   >
@@ -242,8 +268,8 @@ render() {
                     className="form-control"
                     id="repairDescriptionInput"
                     placeholder="mow grass"
-                    value={this.state.eventToSave.title}
-                    name="repairDescription"
+                    value={this.state.title}
+                    name="title"
                     onChange={this.handleInputChange}
                   />
                   </div>
@@ -253,7 +279,7 @@ render() {
                     type="date"
                     className="form-control"
                     id="start-date-input"
-                    value={this.state.eventToSave.startDate}
+                    value={this.state.startDate}
                     name="startDate"
                     onChange={this.handleInputChange}
                     />
@@ -264,7 +290,7 @@ render() {
                     type="date"
                     className="form-control"
                     id="end-date-input"
-                    value={this.state.eventToSave.endDate}
+                    value={this.state.endDate}
                     name="endDate"
                     onChange={this.handleInputChange}
                     />
@@ -275,7 +301,7 @@ render() {
                     type="time"
                     className="form-control"
                     id="start-time-input"
-                    value={this.state.eventToSave.startTime}
+                    value={this.state.startTime}
                     name="startTime"
                     onChange={this.handleInputChange}
                     />                  
@@ -286,7 +312,7 @@ render() {
                     type="time"
                     className="form-control"
                     id="end-time-input"
-                    value={this.state.eventToSave.endTime}
+                    value={this.state.endTime}
                     name="endTime"
                     onChange={this.handleInputChange}
                     />                  
@@ -298,8 +324,8 @@ render() {
                     className="form-control"
                     id="repairCostInput"
                     placeholder="$35"
-                    value={this.state.eventToSave.cost}
-                    name="repairCost"
+                    value={this.state.cost}
+                    name="cost"
                     onChange={this.handleInputChange}
                   />
                 </div>
@@ -308,8 +334,8 @@ render() {
                   <select
                     className="form-control"
                     id="repairPrioritySelect"
-                    value={this.state.eventToSave.priority}
-                    name="repairPriority"
+                    value={this.state.priority}
+                    name="priority"
                     onChange={this.handleInputChange}
                   >
                     <option>low</option>
@@ -323,8 +349,8 @@ render() {
                   <select
                     className="form-control"
                     id="repairStatusSelect"
-                    value={this.state.eventToSave.status}
-                    name="repairStatus"
+                    value={this.state.status}
+                    name="status"
                     onChange={this.handleInputChange}
                   >
                     <option>Thinking about it!</option>
@@ -342,7 +368,7 @@ render() {
                     <div className="col-4">
                       <input
                         type="checkbox"
-                        checked={this.state.eventToSave.isVendor}
+                        checked={this.state.isVendor}
                         onChange={this.handleCheck}
                       />
                     </div>
@@ -355,8 +381,8 @@ render() {
                     <select
                       className="form-control"
                       id="repairVendorSelect"
-                      value={this.state.eventToSave.vendor}
-                      name="repairVendor"
+                      value={this.state.vendor}
+                      name="vendor"
                       onChange={this.handleInputChange}
                     >
                       <option>Vendor 1</option>
@@ -373,8 +399,8 @@ render() {
                     className="form-control"
                     id="repairNotesInput"
                     rows="3"
-                    value={this.state.eventToSave.notes}
-                    name="repairNotes"
+                    value={this.state.notes}
+                    name="notes"
                     onChange={this.handleInputChange}
                   />
                 </div>
@@ -412,7 +438,7 @@ render() {
                   onClick={this.handleFormSubmit}
                   className="btn btn-block btn-success mt-3"
                 >
-                  Add {this.state.eventToSave.repairType}
+                  Add
                 </button>
               </form>
                 </ModalBody>
